Polyester

Polyester.UnsignedIteratorEarlyStop

Polyester.@batch

Polyester.UnsignedIteratorEarlyStop — Type

UnsignedIteratorEarlyStop(thread_mask[, num_threads = count_ones(thread_mask)])

Iterator, returning (i,t) = Tuple{UInt32,UInt32}, where i iterates from 1,2,...,num_threads, and t gives the threadids to call ThreadingUtilities.taskpointer with.

Unfortunately, codegen is suboptimal when used in the ergonomic for (i,tid) ∈ thread_iterator fashion. If you want to microoptimize, You'd get better performance from a pattern like:

function sumk(u,l = count_ones(u) % UInt32)
    uu = ServiceSolicitation.UnsignedIteratorEarlyStop(u,l)
    s = zero(UInt32); state = ServiceSolicitation.initial_state(uu)
    while true
        iter = iterate(uu, state)
        iter === nothing && break
        (i,t),state = iter
        s += t
    end
    s
end

This iterator will iterate at least once; it's important to check and exit early with a single threaded version.

Polyester.@batch — Macro

@batch for i in Iter; ...; end

Evaluate the loop on multiple threads.

@batch minbatch=N for i in Iter; ...; end

Evaluate at least N iterations per thread. Will use at most length(Iter) ÷ N threads.

@batch per=core for i in Iter; ...; end
@batch per=thread for i in Iter; ...; end

Use at most 1 thread per physical core, or 1 thread per CPU thread, respectively. One thread per core will mean less threads competing for the cache, while (for example) if there are two hardware threads per physical core, then using each thread means that there are two independent instruction streams feeding the CPU's execution units. When one of these streams isn't enough to make the most of out of order execution, this could increase total throughput.

Which performs better will depend on the workload, so if you're not sure it may be worth benchmarking both.

LoopVectorization.jl currently only uses up to 1 thread per physical core. Because there is some overhead to switching the number of threads used, per=core is @batch's default, so that Polyester.@batch and LoopVectorization.@tturbo work well together by default.

You can pass both per=(core/thread) and minbatch=N options at the same time, e.g.

@batch per=thread minbatch=2000 for i in Iter; ...; end
@batch minbatch=5000 per=core   for i in Iter; ...; end